Add field TEMP to Appuser table to check if user is temporary. We can delete only temp = true users, other we make used = false

+++ SeedShop/src/main/java/com/apu/seedshop/services/UserService.java
@@ -77,7 +77,10 @@ UserRepository userRepository;
             logger.debug(String.format("Deleting users %s %s %s with id %s", u.getFirstName(), u.getSecName(), u.getThirdName(), u.getUserId()));
             //List<Ugroup> gl = u.getUgroupList();
             //detailsRepository.delete(id);
-            userRepository.delete(id);
+            if(u.getTemp() == true)
+                userRepository.delete(id);
+            else
+                u.setUsed(false);
         }
     }
